// api/model/userRoleResponse.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleDto } from './roleDto';
import { DirectoryDto } from './directoryDto';
import { WorkspaceDto } from './workspaceDto';


export interface UserRoleResponse { 
    id?: string | null;
    role?: RoleDto;
    directoryId?: string | null;
    directory?: DirectoryDto;
    workspaceId?: string | null;
    workspace?: WorkspaceDto;
    userId?: string | null;
}