// api/model/campaignCouponDto.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CouponDto } from './couponDto';


export interface CampaignCouponDto { 
    campaignId?: string | null;
    campaignName?: string | null;
    subTitle?: string | null;
    description?: string | null;
    startDate?: string | null;
    endDate?: string | null;
    coupon?: CouponDto;
    photoUrls?: Array<string> | null;
}